fix(settings): improve validation for PUT methods and enforce consistent ObjectId usage in Settings Module. (#404)
* feat(Settings): Implements Settings Module & User-Anomaly endpoint

This  commit creates settings endpoint with Create, update and read endpoints that are course and user bound.

This commit also adds a Creation endpoint for User-Anomaly.

* fix(settings): improve validation for PUT methods and enforce consistent ObjectId usage

This commit enforces consistent usage of ObjectId and added a custom validator `containsAllDetectors` to ensure that the PUT methods of settings contains all detectors duting updation.

* fix: Settings Modules config changes.

Adds Settings module to tsconfig and package.json